// models/_project.ts
export const projectModel = {
  languages: {
    default: {
      codename: 'default',
      name: 'Default project language',
    },
  },
  contentTypes: {
    block_with_image: {
      codename: 'block_with_image',
      name: 'Block with image',
      elements: {
        title: {
          codename: 'title',
          name: 'Title',
        },
        image: {
          codename: 'image',
          name: 'Image',
        },
        content: {
          codename: 'content',
          name: 'Content',
        },
      },
    },
    component__code_block: {
      codename: 'component__code_block',
      name: 'Component: Code block',
      elements: {
        code: {
          codename: 'code',
          name: 'Code',
        },
        lang: {
          codename: 'lang',
          name: 'Lang',
        },
      },
    },
    content: {
      codename: 'content',
      name: 'Content',
      elements: {
        local_slug: {
          codename: 'local_slug',
          name: 'Local slug',
        },
        content_tags: {
          codename: 'content_tags',
          name: 'Content tags',
        },
        text_3960151: {
          codename: 'text_3960151',
          name: 'Text',
        },
        content_type: {
          codename: 'content_type',
          name: 'Content type',
        },
        url: {
          codename: 'url',
          name: 'URL',
        },
        cover_image: {
          codename: 'cover_image',
          name: 'Cover image',
        },
        cover_image_url: {
          codename: 'cover_image_url',
          name: 'Cover image URL',
        },
        title: {
          codename: 'title',
          name: 'Title',
        },
        teaser: {
          codename: 'teaser',
          name: 'Teaser',
        },
        date: {
          codename: 'date',
          name: 'Date',
        },
      },
    },
    homepage: {
      codename: 'homepage',
      name: 'Homepage',
      elements: {
        subpages: {
          codename: 'subpages',
          name: 'Subpages',
        },
        title: {
          codename: 'title',
          name: 'Title',
        },
        content: {
          codename: 'content',
          name: 'Content',
        },
      },
    },
    page: {
      codename: 'page',
      name: 'Page',
      elements: {
        content: {
          codename: 'content',
          name: 'Content',
        },
      },
    },
  },
  taxonomies: {
    content_tags: {
      codename: 'content_tags',
      name: 'Content tags',
      terms: {
        jamstack: {
          codename: 'jamstack',
          name: 'jamstack',
          terms: {},
        },
        _net: {
          codename: '_net',
          name: '.NET',
          terms: {},
        },
        gatsby: {
          codename: 'gatsby',
          name: 'Gatsby',
          terms: {},
        },
        kentico: {
          codename: 'kentico',
          name: 'Kentico',
          terms: {},
        },
        sourcebit: {
          codename: 'sourcebit',
          name: 'Sourcebit',
          terms: {},
        },
        headless_cms: {
          codename: 'headless_cms',
          name: 'headless CMS',
          terms: {},
        },
        contentful: {
          codename: 'contentful',
          name: 'Contentful',
          terms: {},
        },
        kentico_kontent: {
          codename: 'kentico_kontent',
          name: 'Kentico Kontent',
          terms: {},
        },
        html: {
          codename: 'html',
          name: 'HTML',
          terms: {},
        },
        flexbox: {
          codename: 'flexbox',
          name: 'Flexbox',
          terms: {},
        },
        grid: {
          codename: 'grid',
          name: 'Grid',
          terms: {},
        },
        javascript: {
          codename: 'javascript',
          name: 'JavaScript',
          terms: {},
        },
        unit_test: {
          codename: 'unit_test',
          name: 'unit test',
          terms: {},
        },
        free: {
          codename: 'free',
          name: 'free',
          terms: {},
        },
        gridsome: {
          codename: 'gridsome',
          name: 'Gridsome',
          terms: {},
        },
        git: {
          codename: 'git',
          name: 'git',
          terms: {},
        },
        graph_ql: {
          codename: 'graph_ql',
          name: 'Graph QL',
          terms: {},
        },
        angular: {
          codename: 'angular',
          name: 'Angular',
          terms: {},
        },
      },
    },
    content_type: {
      codename: 'content_type',
      name: 'Content type',
      terms: {
        video: {
          codename: 'video',
          name: 'video',
          terms: {},
        },
        article: {
          codename: 'article',
          name: 'article',
          terms: {},
        },
        live_stream: {
          codename: 'live_stream',
          name: 'live stream',
          terms: {},
        },
      },
    },
  },
};